Some documentation: Cool performance improvement While writing documentation for glvxPaintMask, looking at the calculateZeroCurvaturePoint section, I got the notion that perhaps doing the two "bumps" of a curve as separate drawing operations was unnecessary (meaning I could avoid doing the calculation for the zero curvature point, which would definitely help on performance).
It turns out that I was right.

It looks like what happens is that the triangles drawing in
erroneous space "before" the "maximum" are canceled out exactly
by triangles drawing in the same erroneous space "after" the
"maximum".

It looks like this might only by about a 5% performance improvement
(on a Debug build), but it's still pretty cool nonetheless.

It is also possible that there is some edge case where this does
not work. However, it looks like there is some inherent symmetry
to the thing, meaning that this might not actually be a problem.