Significant change to stroke API; add stroke side colors and offset Stroke properties are now set with glvx(Left|Right|Begin|End)Color, glvx(Begin|End)Width, and glvxStrokeOffset.
The new stroke offset allows strokes to be uncentered.

The new stroke left and right colors are multiplied with interpolated
colors across the stroke in order to change the color across the stroke
normal direction.

The combination of left and right color and stroke offset enables some
new effects, such as curved gradients.