Fix HSV interpolation for stroke; new stroke color model By "new color model" I mean that now, instead of specifying left and right colors that are multiplied with begin and end colors, colors are simply specified for all four corners.
This actually expands the possibilities that can be done with stroke
colors, but does change the API a bit.

Several methods were removed that need to be re-implemented.

HSV interpolation for strokes is especially strange, because strokes
are already doing interpolation for the vertex colors. So the stroke
HSV interpolation has to have a separate shader that does not process
vertex colors.

However, this does mean that there isn't as much unnecessary slow-ness
as there would be otherwise.